feat: 리팩토링 진행중학교 ㅋㅋ
리팩토링을 진행중입니다
