+++ Projects/babyMoa/babyMoa/App/AppDelegate.swift
@@ -8,11 +8,11 @@ import UIKit
 
 class AppDelegate: NSObject, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
-        // UserDefault 데이터 삭제 필요 시
-        if let bundleID = Bundle.main.bundleIdentifier {
-                UserDefaults.standard.removePersistentDomain(forName: bundleID)
-                print("✅ UserDefaults cleared for \(bundleID)")
-            }
+//        // UserDefault 데이터 삭제 필요 시
+//        if let bundleID = Bundle.main.bundleIdentifier {
+//                UserDefaults.standard.removePersistentDomain(forName: bundleID)
+//                print("✅ UserDefaults cleared for \(bundleID)")
+//            }
         return true
     }
 }
