+++ Projects/babyMoa/babyMoa/Models/BabySummary.swift
@@ -1,0 +1,15 @@
+//
+//  BabySummary.swift
+//  babyMoa
+//
+//  Created by 한건희 on 11/2/25.
+//
+
+import UIKit
+
+// TODO: 서버에서 babyId에 대한 아기의 데이터만 내려주는 api 추가 필요
+struct BabySummary {
+    var babyId: String
+    var babyName: String
+    var babyProfileImage: UIImage
+}
