+++ Projects/babyMoa/babyMoa/Utilities/UserDefault/SelectedBaby.swift
@@ -1,0 +1,11 @@
+//
+//  SelectedBaby.swift
+//  babyMoa
+//
+//  Created by 한건희 on 11/2/25.
+//
+
+struct SelectedBaby {
+    @UserDefault(key: "babyId", defaultValue: nil)
+    static var babyId: Int?
+}
