+++ Projects/babyMoa/babyMoa/Views/Growth/SubView/TopBabySummaryView.swift
@@ -1,0 +1,18 @@
+//
+//  TopBabySummaryView.swift
+//  babyMoa
+//
+//  Created by 한건희 on 11/2/25.
+//
+
+import SwiftUI
+
+struct TopBabySummaryView: View {
+    @Binding var viewModel: GrowthViewModel
+    
+    var body: some View {
+        HStack(spacing: 0) {
+//            Image("")
+        }
+    }
+}
